chore: add approve screen
