+++ apps/vesting/src/components/vesting/content/AccountDetails.tsx
@@ -30,7 +30,7 @@ export const AccountDetails = ({ account = "" }: { account?: string }) => {
       return <BannerMessages text={vestingDetails} />;
     }
     const accountName = getVestingAccountNameLocalstorage(
-      vestingDetails.accountAddress
+      vestingDetails.accountAddress,
     );
 
     return (
